server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
beanie==1.25.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.18.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
joblib==1.3.2
lazy-model==0.2.0
motor==3.3.2
numpy==1.26.4
pandas==2.2.1
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.3.0
toml==0.10.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
